+++ data/repository/src/main/java/jp/chau2chaun2/honkot/sample/repository/ImageRepository.kt
@@ -1,11 +1,10 @@
-package jp.chau2chaun2.honkot.sample.multimodule.repository
+package jp.chau2chaun2.honkot.sample.repository
 
 import android.content.Context
 import android.content.res.Resources
 import android.graphics.Bitmap
 import android.graphics.BitmapFactory
 import com.squareup.picasso.Picasso
-import jp.chau2chaun2.honkot.sample.multimodule.R
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import javax.inject.Inject
@@ -18,15 +17,15 @@ class ImageRepository @Inject constructor(
 ) {
     suspend fun loadImage(path: String?): Bitmap? = withContext(Dispatchers.IO) {
         return@withContext if (path?.isEmpty() == true) {
-            BitmapFactory.decodeResource(resources, R.mipmap.ic_launcher)
+            BitmapFactory.decodeResource(resources, R.drawable.ic_none)
         } else {
             try {
                 Picasso.with(context)
                     .load(path)
-                    .error(R.mipmap.ic_launcher)
+                    .error(R.drawable.ic_none)
                     .get()
             } catch (e: Exception) {
-                BitmapFactory.decodeResource(resources, R.mipmap.ic_launcher)
+                BitmapFactory.decodeResource(resources, R.drawable.ic_none)
             }
         }
     }
