+++ settings.gradle
@@ -3,4 +3,5 @@ include ':app'
 include ':data:api'
 include ':data:model'
 include ':data:dao'
+include ':data:repository'
 include ':util'
